css-align-3_dev/html/place-self-shorthand-003.htm

<!DOCTYPE html>
<html><head><title>CSS Box Alignment: place-self shorthand - initial value</title>
<link href="mailto:jfernandez@igalia.com" rel="author" title="Javier Fernandez">
<link href="http://www.w3.org/TR/css3-align/#place-self-property" rel="help">
<meta content="Check that place-self's 'initial' value expands to 'align-self' and 'justify-self'." name="assert">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head><body><div id="log"></div>
<script>
    var div = document.createElement("div");
    document.body.appendChild(div);
    div.style["align-self"] = "start";
    div.style["justify-self"] = "end";
    div.setAttribute("style", "place-self: initial");

    test(function() {
        assert_equals(div.style["align-self"],
            "initial", "place-self specified value for align-self");
    }, "Check place-self: initial - align-self expanded value");

    test(function() {
        assert_equals(div.style["justify-self"],
            "initial", "place-self specified value for justify-self");
    }, "Check place-self: initial - justify-self expanded value");
</script>
</body></html>
